AMEVA-Model-Nexus/requirements.txt
fastapi
uvicorn[standard]
llama-cpp-python
psutil
